Add helper for testing advisory locks
