+++ spec/services/customers/lock_service_spec.rb
@@ -23,35 +23,12 @@ RSpec.describe Customers::LockService do
     end
 
     context "when lock cannot be acquired", transaction: false do
-      let(:lock_released_after) { 2.seconds }
       let(:timeout_seconds) { 0.seconds }
 
       around do |test|
-        customer_id = customer.id
-        queue = Queue.new
-        thread = start_lock_thread(queue, customer_id)
-        sleep 0.5
-        test.run
-      ensure
-        stop_thread(thread, queue) if thread
-      end
-
-      def start_lock_thread(queue, customer_id)
-        Thread.start do
-          start_time = Time.zone.now
-          ApplicationRecord.transaction do
-            ApplicationRecord.with_advisory_lock!("customer-#{customer_id}", transaction: true) do
-              until queue.size > 0 || Time.zone.now - start_time > lock_released_after
-                sleep 0.01
-              end
-            end
-          end
+        with_advisory_lock("customer-#{customer.id}", lock_released_after: 2.seconds) do
+          test.run
         end
-      end
-
-      def stop_thread(thread, queue)
-        queue.push(true)
-        thread.join
       end
 
       it "raises a Customers::FailedToAcquireLock error" do
