+++ spec/spec_helper.rb
@@ -72,6 +72,7 @@ RSpec.configure do |config|
   config.include StripeHelper
   config.include QueuesHelper
   config.include XMLHelper
+  config.include AdvisoryLockHelper
   config.include ActiveSupport::Testing::TimeHelpers
   config.include ActiveStorageValidations::Matchers
   config.include Karafka::Testing::RSpec::Helpers
